Move remaining late app configuration to the normal time
Late configuration was a leftover from when app.config was modified after creating the app, which isn't supported anymore (using real configuration files is the way to go). Late configuration was also causing problems with DB schema creation, as now AuthMethods (and supporting models) aren't imported as they were with the old non-basic data types configuration. By moving the AuthMethod configuration to the rest of the app's setup, AuthMethods' supporting models are now imported and they're registered on db.Model.