create_app can now handle multiple kinds of configuration
Before create_app could only take a filename to use as a configuration source. Now it accepts dictionary instances, strings naming an importable module, or naming a path to a configuration file in the "EVESRP_SETTINGS" environment variable in addition to the existing method of a path to a configuration file.