Do not re-create the entire navbar when updating
While this does remove the functionality where the new buttons appear and disappear when they're added, that situation is rare and can be handled by a user reloading. The common case of the numbers being reloaded is faster and it doesn't rewrite the links, which was causing issues for J4LP as they host on a non-root path.

It also adds a badge for incomplete requests to the user's button.