Add workaround for accessing CREST over TLS
CCP supports a single cipher suite on the public crest endpoint over TLS. This suite uses RC4, which has been removed from the default cipher list for Python. This workaround enables this RC4, but only for that server, and only as a last resort (for when CCP eventually fixes itself).

Requests was also updated to the latest version to take advantage of numerous fixes.