Use CREST for looking up new location data
Again, removes the need to do a new release whenever CCP adds a new system (like they did with the new Wormhole systems). The interface for looking up what constellation a system is in, or the region a constellation is in also changed from using string names to integer IDs.

Also added test cases for the new location and ship lookup process.