Update location relationship lookup method
CREST gained a couple of new fields for some responses, so the System->Constellation got a bit easier. NameLookup was updated to support nested attribute fields, as well as custom attributes to look up (not just the name). Right now Constellation->Region lookups are using a modified version of the old method, as the constellation resources didn’t get as much attention as the region and system resources did. If the region IDs and constellation IDs are added to the constellation resources, that lookup can be simplified down as well. The tests were also updated with the new fields.