Move Sauce Connect tunnel setup into .travis.yml
Comment explaining why it's done in .travis.yml reproduced here:

Sauce Connect setup must be done in .travis.yml as Make uses a fresh shell for each command (line in the file). This means the environment variables needed to find the tunnel are discarded. By doing them here, they are saved in the single session used in Travis.