Switch to using a deploy key for push access to GitHub
Deploy keys are restricted to a single repository instead of every repository I have access to with an OAuth token (well, every personal repository). They're a little more unwieldy to use (SSH key, add it to the agent) but it also means there's basically no chance of they key being exposed in the test log.