Replace bravado with more manual route building
This allows the use of alternate routes (specifically the versioned ones) and now that that’s being done, I can warn when a route has changed.

The initial reason this was done was every instantiation  of a subclass of CcpStore was taking 20 seconds each because it was fetching the Swagger spec each time. This could’ve been solved by caching that request, but it was a good excuse to figure out how to handle deprecation better anyways.