Refine BaseStorage API
add_* methods were reworked to take the arguments to create a model object directly, and return a fully formed one. This feels better than the awkward “create a model with an empty id, add it, then get that id out of the store”.

There’s also more pilot->character changes in this commit.