Ripping out the errors as return values
Thought about it a bit, and decided that not using Python’s exception handling is kinda odd (the not using a language feature over ad-hoc protocols).

It also makes exposing errors up to the end-user a bit easier, as web interfaces will be able to use Sentry, and a CLI interface can handle it itself.