Move all resolving methods to a middleware
Allows for different store types to return custom middlewares, allowing optimized retrieval. For example, an SQL store could have a resolver that peeks at the whole query and fetch all of the objects it needs (and just the columns it needs) in one query instead of multiple queries for each row.