Rewrite search_filter
Took a while to do this, but I think this is a much better approach than the old one. It actually has a more consistent API between sorts and filters, the naming is more consistent between ‘filters’ and ‘searches’ and it should make implementing database-based stores easier with the built-in simplification of filters.