Implement SqlStore modifier voiding
As part of this the void_* columns were removed from the modifier table and a new table, ‘void_modifier’ with the same data was added. This made it easier to distinguish in the SqlStore.void_modifier method between nonexistent modifiers, and those already voided. It also simplified handling the voiding timestamp, as now it’s a server-side default.