Release v0.12.9: OAuth Refresh, ESI and Better Clipboards
OAuth refresh tokens are also used to keep users logged in and to keep their group memberships up to date.

ESI links for killmails are now supported. CREST killmail links are no longer supported, as CREST killmail support is now dead.

ON browsers that support it (most of them now), the JS clipboard API is used. On those that only support Flash, ZeroClipboard is still used. Barring that, a backup option is now available for use as well.

Other bugfixes:
* zKillboard links work again.
* Fix a handful of translation/internationalization bugs.
* Source maps were broken. Now they shouldn’t be.
* Characters not in an alliance will now have nothing shown there instead of “None”